+++ app/Http/Controllers/ImagesController.php
@@ -19,42 +19,41 @@ class imagesController extends Controller
 
     public function store(Request $request)
     {
-        if ($request->hasFile('image')) {
-            //  Let's do everything here
-            if ($request->file('image')->isValid()) {
-                //
-                $validated = $request->validate([
-                    'research_id' => 'required',
-                    'name' => 'unique:images|string|max:40',
-                    'image' => 'required|mimes:jpeg,png,jpg,gif,svg|max:1014',
-                ]);
+        if (Auth::user()->user_type == 'admin') {
+            if ($request->hasFile('image')) {
+                //  Let's do everything here
+                if ($request->file('image')->isValid()) {
+                    //
+                    $validated = $request->validate([
+                        'research_id' => 'required',
+                        'name' => 'unique:images|string|max:40',
+                        'image' => 'required|mimes:jpeg,png,jpg,gif,svg|max:1014',
+                    ]);
+                    $fileName = $validated['name'] . '.' . request()->image->extension();
+                    request()->image->move(public_path('uploads'), $fileName);
 
-                $fileName = $validated['name'] . '.' . request()->image->extension();
-                request()->image->move(public_path('uploads'), $fileName);
-
-                $file = images::create([
-                    'research_id' => $request->research_id,
-                    'name' => $fileName,
-                    'image' => "/uploads/" . $fileName,
-                ]);
-                return redirect()->back();
+                    $file = images::create([
+                        'research_id' => $request->research_id,
+                        'name' => $fileName,
+                        'image' => "/uploads/" . $fileName,
+                    ]);
+                    return redirect()->back()->with('success', 'Image added successfully.');
+                }
+                return redirect()->back()->with('errormsg', 'Invalid image');
             }
-            dd("Image is invalid");
+            return redirect()->back()->with('errormsg', 'Image not recieved.');
         }
-        dd("Image not recieved");
-        return redirect()->back();
-        // abort(500, 'Could not upload image :(');
-        // $extension = $request->image->extension();
-        // $url = "storage/".$validated['name'].".".$extension;
-        // $request->image->storeAs('public/', $validated['name'].".".$extension);
-        // return response()->json(['success'=>'You have successfully upload file.']);
+        return redirect()->back()->with('errormsg', 'You do not have access to delete an image.');
     }
 
     public function destroy($id)
     {
-        $img = Images::where('id', $id)->first();
-        File::delete(public_path() . $img->image);
-        $img->delete();
-        return redirect()->back();
+        if (Auth::user()->user_type == 'admin') {
+            $img = Images::where('id', $id)->first();
+            File::delete(public_path() . $img->image);
+            $img->delete();
+            return redirect()->back()->with('success', 'Image deleted successfully.');
+        }
+        return redirect()->back()->with('errormsg', 'You do not have access to delete an image.');
     }
 }
