+++ app/Http/Controllers/PortfolioController.php
@@ -18,59 +18,46 @@ class portfolioController extends Controller
 
     public function store(Request $request)
     {
-        Validator::make($request->all(), [
-            'company_name' => 'required|max:255',
-            'share_percentage' => 'required|numeric|max:100|min:1',
-            'action' => 'required|max:255',
-            'share_price' => 'required|numeric',
-            'date' => 'required|max:255',
-        ])->validate();
+        if (Auth::user()->user_type == 'admin') {
+            Validator::make($request->all(), [
+                'company_name' => 'required|max:255',
+                'share_percentage' => 'required|numeric|max:100|min:1',
+                'action' => 'required|max:255',
+                'share_price' => 'required|numeric',
+                'date' => 'required|max:255',
+            ])->validate();
 
-        portfolio::create($request->all());
-        return redirect()->back();
+            portfolio::create($request->all());
+            return redirect()->back()->with('success', 'Portfolio added successfully.');
+        }
+        return redirect()->back()->with('errormsg', 'You do not have access to add a portfolio.');
     }
-
-    // public function show($id)
-    // {
-    //     $portfolio = portfolio::where('id', $id)->first();
-    //     if ($portfolio) {
-    //         return view('portfolio.show', compact('portfolio'));
-    //     }
-    //     $error='No portfolio exists with this ID.';
-    //     return redirect()->back()->with('error', $error);
-    // }
-
-    // public function edit($id)
-    // {
-    //     $portfolio = portfolio::where('id', $id)->first();
-    //     // if ($portfolio && (Auth::user()->user_type == 'admin')) {
-    //     if ($portfolio) {
-    //         return view('portfolio.edit', compact('portfolio'));
-    //     } else {
-    //         $error = 'Only Admin can edit the portfolio.';
-    //         return redirect()->back()->with('error',$error);
-    //     }
-    // }
 
     public function update(Request $request, $id)
     {
-        $updatedportfolio = portfolio::where('id', $id)->first();
+        if (Auth::user()->user_type == 'admin') {
+            $updatedportfolio = portfolio::where('id', $id)->first();
 
-        Validator::make($request->all(), [
-            'company_name' => 'required|max:255',
-            'share_percentage' => 'required|numeric|max:255',
-            'action' => 'required|max:255',
-            'share_price' => 'required|numeric',
-            'date' => 'required|max:255',
-        ])->validate();
-        
-        $updatedportfolio->update($request->all());
-        return redirect()->back();
+            Validator::make($request->all(), [
+                'company_name' => 'required|max:255',
+                'share_percentage' => 'required|numeric|max:255',
+                'action' => 'required|max:255',
+                'share_price' => 'required|numeric',
+                'date' => 'required|max:255',
+            ])->validate();
+
+            $updatedportfolio->update($request->all());
+            return redirect()->back()->with('success', 'Portfolio edited successfully.');
+        }
+        return redirect()->back()->with('errormsg', 'You do not have access to edit a portfolio.');
     }
 
     public function destroy($id)
     {
-        portfolio::where('id', $id)->first()->delete();
-        return redirect()->back();
+        if (Auth::user()->user_type == 'admin') {
+            portfolio::where('id', $id)->first()->delete();
+            return redirect()->back()->with('success', 'Portfolio deleted successfully.');
+        }
+        return redirect()->back()->with('errormsg', 'You do not have access to edit a portfolio.');
     }
 }
